- 개인정보수정 API 개발완료 - 게시글 등록, 수정, 삭제, 검색, 리스트 조회, 상세 조회 API 개발완료

+++ user/src/main/java/com/hsj/aft/user/dto/UserDto.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 public class UserDto {
 
-    private Long userNo;
+    private int userNo;
     private String userId;
     private String userName;
