+++ user/src/main/java/com/hsj/aft/user/repository/UserRepository.java
@@ -3,8 +3,6 @@ package com.hsj.aft.user.repository;
 import com.hsj.aft.domain.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Long> {
-
-
+public interface UserRepository extends JpaRepository<User, Integer> {
 
 }
